fix signed/unsigned type errors in game_engine_has_shield and FUN_0005c940
game_engine_has_shield (0xa95f0, game_engine.c):
  *(int *)0x456b18 >> 3  →  *(unsigned int *)0x456b18 >> 3
  Signed right shift emitted sarl; original binary uses shrl.
  VC71 match: 92.9% → 100.0%.

FUN_0005c940 (0x5c940, encounters.c):
  (int)(short)(encounter_handle & 0xffff)  →  (encounter_handle & 0xffff)
  The (short) cast emitted movswl (sign-extend); original binary zero-extends.
  Also a correctness fix: the oracle is unsigned. VC71 match: 92.9% → 95.1%.

+++ src/halo/ai/encounters.c
@@ -3798,7 +3798,7 @@ void FUN_0005c940(int encounter_handle)
   encounter = (char *)datum_get(*(data_t **)0x5ab270, encounter_handle);
   enc_def_elt = (char *)tag_block_get_element(
     (char *)global_scenario_get() + 0x42c,
-    (int)(short)(encounter_handle & 0xffff), 0xb0);
+    (encounter_handle & 0xffff), 0xb0);
   i = 0;
   if (*(short *)(encounter + 0xa) <= 0) {
     return;
